libaqr: rename libaqr to libaqua
