+++ tools/libaqua/errors.py
